PR changes and merge into mqtt5
