+++ include/aws/mqtt/v5/mqtt5_types.h
@@ -140,6 +140,17 @@ enum aws_mqtt5_puback_reason_code {
 };
 
 /**
+ * Result for manual PUBACK operations.
+ *
+ */
+enum aws_mqtt5_manual_puback_result {
+    AWS_MQTT5_MPR_SUCCESS = 0,
+    AWS_MQTT5_MPR_PUBACK_CANCELLED = 1,
+    AWS_MQTT5_MPR_PUBACK_INVALID = 2,
+    AWS_MQTT5_MPR_CRT_FAILURE = 3,
+};
+
+/**
  * Reason code inside SUBACK packet payloads.
  * Enum values match mqtt spec encoding values.
  *
@@ -441,6 +452,18 @@ struct aws_mqtt5_packet_puback_view {
 
     size_t user_property_count;
     const struct aws_mqtt5_user_property *user_properties;
+};
+
+/**
+ * This is used to track which PUBLISH packets a user has taken manual PUBACK control from.
+ */
+struct aws_mqtt5_manual_puback_entry {
+    struct aws_allocator *allocator;
+
+    /* control id for internal tracking */
+    uint64_t puback_control_id;
+    /* packet_id of controlled publish */
+    uint16_t packet_id;
 };
 
 /**
